documenting sources. cleaning code
